youtube-translator/pyproject.toml
[project]
name = "youtube-translator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-cloud-speech>=2.36.0",
    "google-cloud-texttospeech>=2.34.0",
    "google-cloud-translate>=3.24.0",
    "pydub>=0.25.1",
    "yt-dlp>=2025.12.8",
]
